Non-Blocking SPSC Queue
Add a non-blocking (non-locking) single-producer, single-consumer queue to Rabit.  This gives a faster queue to Rabit for some use-cases.  The queue suports a single manger suppling items to the queue and a sincle manager extracting items from the queue.  It will work with raw and shared pointers. 

The standard message queue has some cleanup so avoid multiple locks when pushing or popping items.
A change was made to the publish-subscibe message to avoid locking the messege if the time stamp has not changed.